Refactor of zwave value_changed
