+++ homeassistant/components/light/zwave.py
@@ -90,9 +90,6 @@ class ZwaveDimmer(zwave.ZWaveDeviceEntity, Light):
 
     def __init__(self, value, refresh, delay):
         """Initialize the light."""
-        from openzwave.network import ZWaveNetwork
-        from pydispatch import dispatcher
-
         zwave.ZWaveDeviceEntity.__init__(self, value, DOMAIN)
         self._brightness = None
         self._state = None
@@ -118,38 +115,33 @@ class ZwaveDimmer(zwave.ZWaveDeviceEntity, Light):
         self._timer = None
         _LOGGER.debug('self._refreshing=%s self.delay=%s',
                       self._refresh_value, self._delay)
-        dispatcher.connect(
-            self._value_changed, ZWaveNetwork.SIGNAL_VALUE_CHANGED)
 
     def update_properties(self):
         """Update internal properties based on zwave values."""
         # Brightness
         self._brightness, self._state = brightness_state(self._value)
 
-    def _value_changed(self, value):
-        """Called when a value has changed on the network."""
-        if self._value.value_id == value.value_id or \
-           self._value.node == value.node:
-            _LOGGER.debug('Value changed for label %s', self._value.label)
-            if self._refresh_value:
-                if self._refreshing:
-                    self._refreshing = False
-                    self.update_properties()
-                else:
-                    def _refresh_value():
-                        """Used timer callback for delayed value refresh."""
-                        self._refreshing = True
-                        self._value.refresh()
-
-                    if self._timer is not None and self._timer.isAlive():
-                        self._timer.cancel()
-
-                    self._timer = Timer(self._delay, _refresh_value)
-                    self._timer.start()
-                self.schedule_update_ha_state()
-            else:
+    def value_changed(self, value):
+        """Called when a value for this entity's node has changed."""
+        if self._refresh_value:
+            if self._refreshing:
+                self._refreshing = False
                 self.update_properties()
-                self.schedule_update_ha_state()
+            else:
+                def _refresh_value():
+                    """Used timer callback for delayed value refresh."""
+                    self._refreshing = True
+                    self._value.refresh()
+
+                if self._timer is not None and self._timer.isAlive():
+                    self._timer.cancel()
+
+                self._timer = Timer(self._delay, _refresh_value)
+                self._timer.start()
+            self.schedule_update_ha_state()
+        else:
+            self.update_properties()
+            self.schedule_update_ha_state()
 
     @property
     def brightness(self):
